+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Use pageTitle and description in `SearchWrapper` when there is no `searchQuery`
+
 ## [2.111.1] - 2021-02-04
 
 ### Fixed